add jwt and rbac changes

+++ routes/teachers_routes.py
@@ -69,7 +69,9 @@
 
 
 from flask import Blueprint,jsonify,request
-from flask_jwt_extended import create_access_token
+from flask_jwt_extended import create_access_token,create_refresh_token,jwt_required,get_jwt_identity
+from werkzeug.security import generate_password_hash,check_password_hash
+from flask_jwt_extended import get_jwt
 
 from project3.models.teacher_model import Teacher
 from project3.services import teachers_services
@@ -82,14 +84,28 @@ def login():
     teacher=Teacher.query.filter_by(email=email).first()
     if not teacher:
         return jsonify({"error":"teacher not found"}),404
-    if teacher.password!=data["password"]:
+    if not check_password_hash(teacher.password,data["password"]):
         return jsonify({"message":"incorrect password"}),401
-    token=create_access_token(identity=str(teacher.id))
-    return jsonify({"message":"login successfully","token":token}),200
+    access_token=create_access_token(identity=str(teacher.id),
+                                     additional_claims={"role":teacher.role})
+    refresh_token=create_refresh_token(identity=str(teacher.id))
+    return jsonify({"message":"login successfully","access_token":access_token,"refresh_token":refresh_token}),200
+
+@teachers_routes.route("/refresh",methods=["POST"])
+@jwt_required(refresh=True)
+def refresh():
+    teacher_id=get_jwt_identity()
+    teacher=Teacher.query.get(teacher_id)
+    if not teacher:
+        return jsonify({"error":"teacher not found"}),404
+    new_access_token=create_access_token(identity=str(teacher.id),
+                                         additional_claims={"role":teacher.role})
+    return jsonify({"new_access_token":new_access_token}),200
 
 @teachers_routes.route("/",methods=["POST"])
 def add():
     data=request.get_json()
+    data["password"]=generate_password_hash(data["password"])
     error=validate_teacher(data)
     if error:
         return jsonify({"error":error}),400
@@ -98,17 +114,23 @@ def add():
         return jsonify({"error":"email already exists"}),400
     return jsonify(teacher.to_dict()),201
 @teachers_routes.route("/",methods=["GET"])
+@jwt_required()
 def getall():
     teachers=teachers_services.get_all_teachers()
     return jsonify([teacher.to_dict() for teacher in teachers]),200
 @teachers_routes.route("/<int:tid>",methods=["GET"])
+@jwt_required()
 def getbyid(tid):
     teacher=teachers_services.get_teacher_by_id(tid)
     if isinstance(teacher,tuple):
         return jsonify(teacher[0],teacher[1])
     return jsonify(teacher.to_dict()),200
 @teachers_routes.route("/<int:tid>",methods=["PUT"])
+@jwt_required()
 def updatebyid(tid):
+    claims=get_jwt()
+    if claims.get("role")!="admin":
+        return jsonify({"message":"admin access required"}),403
     data=request.get_json()
     error=validate_update_teacher(data)
     if error:
@@ -118,7 +140,11 @@ def updatebyid(tid):
         return jsonify(teacher[0],teacher[1])
     return jsonify(teacher.to_dict()),200
 @teachers_routes.route("/<int:tid>",methods=["DELETE"])
+@jwt_required()
 def delete(tid):
+    claims=get_jwt()
+    if claims.get("role")!="admin":
+        return jsonify({"message":'admin access required'}),403
     teacher=teachers_services.delete_teacher(tid)
     if isinstance(teacher,tuple):
         return jsonify(teacher[0],teacher[1])
